requirement.txt
Flask
numpy
opencv-python
tensorflow
keras
groq
python-dotenv
Flask-Login